Weather station errors
Moved binary header variables 'seed laser temp' and 'power transmitted' out of the weather station WS routine.  These values were not being read or saved when the WS flag was turned off.  May need to add a trap for when the WS data is missing or zero (recent Labview changes which dropped the WS from the binary data header).